move template data from table to the tamplate-overview page
