+++ packages/plugins/auto-doc/src/components/data-table/data-table.svelte
@@ -3,96 +3,50 @@
     import Checkbox from '@smui/checkbox';
     import  IconWrapper from '../../../../../ui/src/components/icons/icon-wrapper.svelte'
     import IconButton from '@smui/icon-button'
+    import { type Template} from "@/pages/template-overview/template-overview.svelte"
 
 
-    type Template = {
-        name: string,
-        description: string,
-        lastEdited: Date,       
-    }
-  
-    let allTemplates : Template[] = [
-      {
-        name: 'Template1',
-        description: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit. Assumenda, asperiores accusamus quis sapiente sequi facilis?',
-        lastEdited: new Date("2024-10-15T08:38:00"),
-      },
-      {
-        name: 'Template256',
-        description: 'A simple description',
-        lastEdited: new Date("2024-10-10T13:24:00")
-      },
-      {
-        name: 'CYK_Template',
-        description: 'I have an invalid Date ',
-        lastEdited: new Date("foo"),
-      },
-      {
-        name: 'I surely have quite a long title name.',
-        description: "Lorem ipsum dolor sit amet consectetur.",
-        lastEdited: new Date("2024-10-20T10:55:32"),
-      },
-      {
-        name: 'AUOBHBT Template',
-        description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Soluta placeat itaque minus praesentium sequi deleniti voluptatibus dicta!',
-        lastEdited: new Date(),
-      },
-    ];
+    export let allTemplates : Template[];
+
     let selectedTemplates : Template[] = [];
 
-    $: totalSelected = selectedTemplates.length
 
-function isInvalidDate(date: Date){
-    return isNaN(date.getTime());
-}
+    function isInvalidDate(date: Date){
+        return isNaN(date.getTime());
+    }
 
-function formatDate(date: Date, defaultString: string = "-"): string{
-    if(isInvalidDate(date)){ return defaultString;}
+    function formatDate(date: Date, defaultString: string = "-"): string{
+        if(isInvalidDate(date)){ return defaultString;}
 
-    return `${getDD_MM_YYYYFromDate(date)}, ${getHH_MM_FromDate(date)}`;
-}
+        return `${getDD_MM_YYYYFromDate(date)}, ${getHH_MM_FromDate(date)}`;
+    }
 
-function getHH_MM_FromDate(date: Date,): string{
+    function getHH_MM_FromDate(date: Date,): string{
 
-    const timeOptions: Intl.DateTimeFormatOptions = { 
-        hour: "numeric",
-        minute: "2-digit"
-    };
+        const timeOptions: Intl.DateTimeFormatOptions = { 
+            hour: "numeric",
+            minute: "2-digit"
+        };
 
-    const extractedHH_MM = new Intl.DateTimeFormat("en-US", timeOptions).format(date)
-    return extractedHH_MM;
-}
-function getDD_MM_YYYYFromDate(date: Date): string{
-    const dateOptions: Intl.DateTimeFormatOptions = { 
-        month:   "2-digit" as const, 
-        day:     "2-digit" as const,
-        year:    "numeric" as const,
-    };
+        const extractedHH_MM = new Intl.DateTimeFormat("en-US", timeOptions).format(date)
+        return extractedHH_MM;
+    }
+    function getDD_MM_YYYYFromDate(date: Date): string{
+        const dateOptions: Intl.DateTimeFormatOptions = { 
+            month:   "2-digit" as const, 
+            day:     "2-digit" as const,
+            year:    "numeric" as const,
+        };
 
-    const DD_MM_YYYY = new Date(date).toLocaleDateString("de-DE", dateOptions); 
-    return DD_MM_YYYY;
-}
+        const DD_MM_YYYY = new Date(date).toLocaleDateString("de-DE", dateOptions); 
+        return DD_MM_YYYY;
+    }
 
-
-
-  </script>
+</script>
 
 
-<header class="template-controls">
-    <button class="btn-pill btn-pill-primary"> + Add template</button>
 
-    <button class="btn-pill btn-pill-outlined">Export Documents</button>
-    <button class="btn-pill btn-pill-outlined">Filter</button>
-
-    <span class="total-selected">
-        {#if selectedTemplates.length > 0}
-            {totalSelected} items selected
-        {/if}
-    </span>
-</header>
-
-
-<main class="table-container">      
+<div class="table-container">      
     <DataTable>
         <Head>
             <Row>
@@ -109,11 +63,11 @@ function getDD_MM_YYYYFromDate(date: Date): string{
             {#each allTemplates as template (template.name)}
             <Row>
                 <Cell checkbox>
-                <Checkbox
-                    bind:group={selectedTemplates}
-                    value={template}
-                    valueKey={template.name}
-                />
+                    <Checkbox
+                        bind:group={selectedTemplates}
+                        value={template}
+                        valueKey={template.name}
+                    />
                 </Cell>
                 <Cell>{template.name}</Cell>
                 <Cell>{formatDate(template.lastEdited)}</Cell>
@@ -138,18 +92,9 @@ function getDD_MM_YYYYFromDate(date: Date): string{
             {/each}
         </Body>
     </DataTable>
-</main>
-
-
-
-  
-
-  
+</div>
 
 <style lang="scss">
-    $clr-purple: #6C71C3;
-    $clr-purple-15: #494fbf;
-
     .total-selected{
         min-height: 1.25rem;
     }
@@ -173,42 +118,9 @@ function getDD_MM_YYYYFromDate(date: Date): string{
             border-radius: 10px;
          }
     }
-
-    .template-controls{
-        margin: 1rem 0 1rem 1rem;
-        .btn-pill{
-            padding: 0.5em 1em;
-            border-radius: 1em;
-            cursor: pointer;
-        }
-
-        .btn-pill-outlined{
-            background-color: transparent;
-            border-color: $clr-purple;
-            color: $clr-purple;
-
-
-            &:hover{
-                background-color: $clr-purple;
-                color: white;
-            }
-        }
-
-        .btn-pill-primary{
-            background-color: $clr-purple;
-            color: white;
-
-            &:hover{
-                background-color:$clr-purple-15;
-                color: white;
-            }
-        }
-
-
-        span{
-            display: inline-block;
-            margin-left: inherit;
-        }
+    span{
+        display: inline-block;
+        margin-left: inherit;
     }
    
 </style>
